Move Netlify function tests out of functions dir to fix deploy
Test files in netlify/functions/ were being bundled as serverless
functions by Netlify, breaking deploy previews. Moved them to
src/netlify-functions/ with updated import paths and ESLint config.
