+++ netlify.toml
@@ -2,10 +2,6 @@
   command = "npm run build"
   publish = "dist"
 
-[functions]
-  directory = "netlify/functions"
-  excluded_files = ["**/*.test.js"]
-
 [[redirects]]
   from = "/*"
   to = "/index.html"
